Average rating, and can't rate the same toilet twice

+++ app/src/main/java/com/example/toilets_finder/ToiletDataStore.kt
@@ -9,7 +9,7 @@ data class Toilet(
     val imageSrc: Int,
     val openingHours: String,
     val pmrAccess: String,
-    val averageRating: Float,
+    val averageRating: Double,
     val yourRating: Float,
     val ficheUrl: String
 )
